Fixing top bar responsiveness
