+++ src/app/locales/en.json
@@ -5,7 +5,7 @@
     "logout": "Logout",
     "signup": "Sign up",
     "signinMsg": "Or sign in ",
-    "signinLink": "here"
+    "signinLink": "Signin"
   },
   "navBar": {
     "collections": "Collections",
