+++ src/presentation/components/TopBar/TopBar.module.scss
@@ -32,6 +32,10 @@
       &Cards {
         margin-right: 3rem;
       }
+
+      &Link {
+        padding: 1rem;
+      }
     }
 
     &Link {
